move app type to common
