+++ kpops/components/common/app_type.py
