+++ kpops/components/streams_bootstrap/producer/producer_app.py
@@ -8,13 +8,13 @@ from kpops.components.base_components.kafka_app import (
     KafkaApp,
     KafkaAppCleaner,
 )
+from kpops.components.common.app_type import AppType
 from kpops.components.common.streams_bootstrap import StreamsBootstrap
 from kpops.components.common.topic import (
     KafkaTopic,
     OutputTopicTypes,
     TopicConfig,
 )
-from kpops.components.streams_bootstrap.app_type import AppType
 from kpops.components.streams_bootstrap.producer.model import ProducerAppValues
 from kpops.utils.docstring import describe_attr
