--- kpops/components/streams_bootstrap_v3/app_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class AppType(Enum):
-    STREAMS_APP = "streams-app"
-    PRODUCER_APP = "producer-app"
-    CLEANUP_STREAMS_APP = "streams-app-cleanup-job"
-    CLEANUP_PRODUCER_APP = "producer-app-cleanup-job"
