+++ kpops/components/streams_bootstrap_v3/streams/streams_app.py
@@ -7,8 +7,8 @@ from typing_extensions import override
 from kpops.component_handlers.kubernetes.pvc_handler import PVCHandler
 from kpops.components.base_components.helm_app import HelmApp
 from kpops.components.base_components.kafka_app import KafkaAppCleaner
+from kpops.components.common.app_type import AppType
 from kpops.components.common.topic import KafkaTopic
-from kpops.components.streams_bootstrap_v3.app_type import AppType
 from kpops.components.streams_bootstrap_v3.base import (
     StreamsBootstrapV3,
 )
